Before Part 89

+++ app/routers/post.py
@@ -1,24 +1,34 @@
-from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi import APIRouter, Depends, Security, status, HTTPException
+from sqlalchemy import func
 from sqlalchemy.orm import Session
-from typing import List
+from typing import List, Optional
 from fastapi.responses import Response
 from ..database import get_db
-from .. import schemas, models
+from .. import schemas, models, oauth2
 
 router = APIRouter(
     prefix="/posts",
-    tags=["Posts"]
+    tags=["Posts"],
+    # dependencies=[Security(oauth2.get_current_user)]
 )
 
-@router.get("/", response_model=list[schemas.PostResponse])
-async def get_posts(db: Session = Depends(get_db)): 
-    my_posts = db.query(models.Post).all()
+
+@router.get("/", response_model=List[schemas.PostResponse])
+async def get_posts(db: Session = Depends(get_db),
+                    user: id = Depends(oauth2.get_current_user),
+                    limit: int = 10, skip: int = 0, search: Optional[str] = ""
+                    ): 
+    my_posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
+    
     return my_posts
 
 
 @router.post("/", response_model=schemas.PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_posts(post : schemas.PostCreate, db: Session = Depends(get_db)):
-    new_post = models.Post(**post.dict())
+async def create_posts(post : schemas.PostCreate, 
+                       db: Session = Depends(get_db), 
+                       user: id = Depends(oauth2.get_current_user)):
+    
+    new_post = models.Post(owner_id = user.id, **post.dict())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -28,7 +38,10 @@ async def create_posts(post : schemas.PostCreate, db: Session = Depends(get_db))
 
 
 @router.get("/{id}", response_model=schemas.PostResponse)
-async def get_post(id: int, db: Session = Depends(get_db)):
+async def get_post(id: int, 
+                   db: Session = Depends(get_db),
+                   user: id = Depends(oauth2.get_current_user)):
+    
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -37,26 +50,40 @@ async def get_post(id: int, db: Session = Depends(get_db)):
     return post
 
 @router.delete("/{id}", status_code = status.HTTP_204_NO_CONTENT)
-async def delete_post(id: int, db: Session = Depends(get_db)):
+async def delete_post(id: int, 
+                      db: Session = Depends(get_db),
+                      user: id = Depends(oauth2.get_current_user)):
+    
     post_query = post = db.query(models.Post).filter(models.Post.id == id)
 
     if not post_query.first():
            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id:{id} does not exist")
     
+    if post_query.first().owner_id != user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to perform the requested action")
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 @router.put("/{id}", response_model=schemas.PostResponse)
-async def update_post(id: int, post: schemas.PostUpdate, db: Session = Depends(get_db)):
+async def update_post(id: int, 
+                      post: schemas.PostUpdate, 
+                      db: Session = Depends(get_db),
+                      user: id = Depends(oauth2.get_current_user)):
+    
+    print(user.id)
     post_query = db.query(models.Post).filter(models.Post.id == id)
 
     if not post_query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id:{id} does not exist")
     
+    if models.Post.owner_id != user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to perform the requested action")
     post_query.update(post.dict(), synchronize_session=False)
     db.commit()
     updatedPost = post_query.first()
